pyproject.toml
[project]
name = "pinkponk"
version = "0.1.0"
description = "inbox zero is a pink feeling"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["httpx==0.28.*"]

[dependency-groups]
dev = ["nest-asyncio>=1.6.0", "pytest-asyncio>=0.25.2", "tryceratops>=2.4.1"]

[tool.pytest.ini_options]
addopts = "--strict-markers --strict --quiet"
asyncio_mode = "strict"
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:distutils.*:",
    "ignore::DeprecationWarning:pkg_resources.*:",
]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
required_plugins = ["pytest-asyncio"]
pythonpath = [".", "src"]
xfail_strict = true
